refactor package name from pyYandexLinguistics to pyLinguist
