+++ pyLinguist/Dictionary.py
