+++ pyLinguist/config.py
